Service updated and deleted updated with validation by doctor id

+++ src/app/modules/Doctor/Doctor.service.ts
@@ -1,19 +1,38 @@
 import httpStatus from 'http-status';
 import AppError from '../../errors/appError';
 import { TService } from './Doctor.interface';
-import { Service } from './Doctor.model';
+import { Doctor, Service } from './Doctor.model';
+import { Types } from 'mongoose';
 
-const createService = async (payload: TService) => {
+const createService = async (payload: TService, userId: string) => {
+  const doctor = await Doctor.findOne({ user: userId });
+
+  if (!doctor) {
+    throw new AppError(httpStatus.NOT_FOUND, 'Doctor not found!');
+  }
+
+  if (doctor) {
+    payload.doctor = doctor._id as unknown as Types.ObjectId;
+  }
+
   const result = await Service.create(payload);
   return result;
 };
 
-const updateService = async (payload: Partial<TService>, id: string) => {
-  const service = await Service.findById(id);
+const updateService = async (
+  payload: Partial<TService>,
+  id: string,
+  userId: string,
+) => {
+  // Find DoctorId
+  const doctorId = await Doctor.findOne({ user: userId }).select('_id');
 
-  // If service not found then throw an error
-  if (!service) {
-    throw new AppError(httpStatus.NOT_FOUND, 'Service not found!');
+  // Find doctorId from service
+  const doctorIdFromService = await Service.findById(id).select('doctor');
+
+  // Compare doctorId from token to doctorIdFromService. To check service is created by this doctor or not.
+  if (String(doctorId?._id) !== String(doctorIdFromService?.doctor)) {
+    throw new AppError(httpStatus.UNAUTHORIZED, 'You are not authorized!');
   }
 
   const result = await Service.findByIdAndUpdate(id, payload, { new: true });
@@ -21,7 +40,18 @@ const updateService = async (payload: Partial<TService>, id: string) => {
   return result;
 };
 
-const deleteService = async (id: string) => {
+const deleteService = async (id: string, userId: string) => {
+  // Find DoctorId
+  const doctorId = await Doctor.findOne({ user: userId }).select('_id');
+
+  // Find doctorId from service
+  const doctorIdFromService = await Service.findById(id).select('doctor');
+
+  // Compare doctorId from token to doctorIdFromService. To check service is created by this doctor or not.
+  if (String(doctorId?._id) !== String(doctorIdFromService?.doctor)) {
+    throw new AppError(httpStatus.UNAUTHORIZED, 'You are not authorized!');
+  }
+
   const result = await Service.findByIdAndDelete(id, { new: true });
   return result;
 };
